Move nsfw to dependencies, remove napi & murmur2 binaries. (#1446)
* Move nsfw to dependencies, remove tmp napi binaries.

* Some fixes

* Downgrading nsfw package to working version

* Add node-gyp on latest version for Windows build

* Revert "Add node-gyp on latest version for Windows build"

This reverts commit 0488609a05a8c433d9ea7cacb564814177597049.

* Add workflow to install node-gyp

* Some debugging, tryouts...

* Some more debug stuff

* Where node plz

* Where node the second

* Yet another try to fix it

* Debuggin & Windoof

* Can't find words to describe my feelings

* I can't believe it worked

* Yet another tryyy

* A bit less upgrade
