+++ scripts/moveNapi.js
@@ -1,5 +1,7 @@
 const fs = require('fs');
-const os = require('os');
+// const os = require('os');
 
-fs.renameSync('./napi.node', `../public/native/${os.platform()}/napi.node`);
+if (!fs.existsSync('../public/native')) fs.mkdirSync('../public/native');
+
+fs.renameSync('./napi.node', `../public/native/napi.node`);
 // fs.renameSync('./index.d.ts', `../public/native/${os.platform()}/index.d.ts`);
